+++ src/Resources/Wacc/index.ts
@@ -6,7 +6,8 @@ import WACCUsers from './Users/Users';
 import WACCUserSubscriptions from './UserSubscriptions/UserSubscriptions';
 import WACCGifts from './Gifts/Gifts';
 import WACCProducts from './Products/Products';
-import WACCStock from './Stock/Toybox';
+import WACCStockToybox from './Stock/Toybox';
+import WACCStockTransition from './Stock/Transition';
 import WACCProductAssociations from './ProductAssociations/ProductAssociations';
 import WACCAssets from './Assets/Assets';
 import WACCReturnOrderLines from './ReturnOrderLines/ReturnOrderLines';
@@ -28,7 +29,8 @@ export interface WaccResources {
     products: WACCProducts;
     productAssociations: WACCProductAssociations;
     assets: WACCAssets;
-    stock: WACCStock;
+    stock: WACCStockToybox;
+    stockTransitions: WACCStockTransition;
     roles: WACCRoles;
     waitSpots: WACCWaitSpots;
     referrals: WACCReferral;
@@ -48,7 +50,8 @@ export function loadWaccResources(client: Client): WaccResources {
         products: new WACCProducts(client),
         productAssociations: new WACCProductAssociations(client),
         assets: new WACCAssets(client),
-        stock: new WACCStock(client),
+        stock: new WACCStockToybox(client),
+        stockTransitions: new WACCStockTransition(client),
         roles: new WACCRoles(client),
         waitSpots: new WACCWaitSpots(client),
         referrals: new WACCReferral(client),