+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 🏡 Features
+- Add Stock Replenishment - Bin, Stock, StockCheck & ProductVariant updates ()
 
 ## [0.0.63] - 2021-07-23
 ### 🐛 Bug fixes